refactoring(@cubejs-backend/api-gateway): Extract query validations to own module
